import: Verify exported realm's migration compatibility.
When transferring a realm to a server that has a different set of
applied migrations (different Zulip versions), there is a chance that
the imported data formats appear to be compatible but data invariants
could still be violated.

This commit adds an assertion during the import process to verify
that both the exported realm and the importing server have matching
Zulip versions and have a compatible set of migrations.

A wrapper function, `import_realm` has been added to the
`ImportExportFile` class (previously called `ExportFile`) to handle the
need for patching all `do_import_realm` calls during testing. The class
has been renamed to better reflect these changes.